backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
crewai>=0.28.8
crewai-tools>=0.1.6
openai>=1.6.1
supabase>=2.3.0
vecs>=0.4.0
pandas>=2.1.4
numpy>=1.26.0
PyPDF2>=3.0.1
XlsxWriter>=0.5.7
python-pptx>=0.6.23
docx2txt>=0.8
tiktoken>=0.5.2
google-api-python-client>=2.110.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.2.0
httpx>=0.25.2
aiofiles>=23.2.0
jinja2>=3.1.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 